tsconfig: move CompilerOptions to models/
